friendsListProvider에 있던 timeAgo 함수를 util 폴더 formattor.dart로 분리
