+++ lib/feature/friend/provider/FriendsListProvider.dart
@@ -4,12 +4,17 @@ import 'package:firebase_storage/firebase_storage.dart';
 
 import '../../../model/FriendModel.dart';
 
-
 // 친구 목록을 제공하는 Provider
 final friendsListProvider = StreamProvider<List<FriendModel>>((ref) {
   final firestore = FirebaseFirestore.instance;
   return firestore.collection('friends').snapshots().map((snapshot) {
-    return snapshot.docs.map((doc) => FriendModel.fromDocument(doc)).toList();
+    return snapshot.docs.map((doc) {
+      final data = doc.data();
+      return FriendModel.fromJson({
+        ...data,
+        'lastConnect': (data['lastConnect'] as Timestamp).toDate().toIso8601String(),
+      });
+    }).toList();
   });
 });
 
@@ -25,16 +30,3 @@ final imageProvider = FutureProvider.family<String, String>((ref, imageName) asy
   return await fetchFriendImageUrl(imageName);
 });
 
-//마지막 접속 시간을 일,시,분 별로 포멧팅해서 제공 해주는 함수
-String timeAgo(DateTime dateTime) {
-  final duration = DateTime.now().difference(dateTime);
-  if (duration.inDays > 1) {
-    return '${duration.inDays} days ago';
-  } else if (duration.inHours > 1) {
-    return '${duration.inHours} hours ago';
-  } else if (duration.inMinutes > 1) {
-    return '${duration.inMinutes} minutes ago';
-  } else {
-    return 'Just now';
-  }
-}
